+++ packages/hardhat-core/src/internal/cli/project-creation.ts
@@ -20,7 +20,7 @@ import {
   confirmProjectCreation,
 } from "./prompt";
 import { emoji } from "./emoji";
-import { Dependencies } from "./types";
+import { Dependencies, PackageManager } from "./types";
 
 enum Action {
   CREATE_JAVASCRIPT_PROJECT_ACTION = "Create a JavaScript project",
@@ -60,7 +60,7 @@ const TYPESCRIPT_DEPENDENCIES: Dependencies = {};
 
 const TYPESCRIPT_PEER_DEPENDENCIES: Dependencies = {
   "@types/chai": "^4.2.0",
-  "@types/mocha": "^9.1.0",
+  "@types/mocha": ">=9.1.0",
   "@types/node": ">=12.0.0",
   "ts-node": ">=8.0.0",
   typescript: ">=4.5.0",
@@ -252,20 +252,6 @@ function showStarOnGitHubMessage() {
   console.log(chalk.cyan("     https://github.com/NomicFoundation/hardhat"));
 }
 
-export function showSoliditySurveyMessage() {
-  if (new Date() > new Date("2023-07-01 23:39")) {
-    // the survey has finished
-    return;
-  }
-
-  console.log();
-  console.log(
-    chalk.cyan(
-      "Please take a moment to complete the 2022 Solidity Survey: https://hardhat.org/solidity-survey-2022"
-    )
-  );
-}
-
 export async function createProject() {
   printAsciiLogo();
 
@@ -301,7 +287,6 @@ export async function createProject() {
 
     console.log();
     showStarOnGitHubMessage();
-    showSoliditySurveyMessage();
 
     return;
   }
@@ -372,7 +357,7 @@ export async function createProject() {
         useDefaultPromptResponses ||
         (await confirmRecommendedDepsInstallation(
           dependenciesToInstall,
-          await isYarnProject()
+          await getProjectPackageManager()
         ));
       if (shouldInstall) {
         const installed = await installRecommendedDependencies(
@@ -402,7 +387,6 @@ export async function createProject() {
   console.log("See the README.md file for some example tasks you can run");
   console.log();
   showStarOnGitHubMessage();
-  showSoliditySurveyMessage();
 }
 
 async function canInstallRecommendedDeps() {
@@ -428,6 +412,16 @@ function isInstalled(dep: string) {
 
 async function isYarnProject() {
   return fsExtra.pathExists("yarn.lock");
+}
+
+async function isPnpmProject() {
+  return fsExtra.pathExists("pnpm-lock.yaml");
+}
+
+async function getProjectPackageManager(): Promise<PackageManager> {
+  if (await isYarnProject()) return "yarn";
+  if (await isPnpmProject()) return "pnpm";
+  return "npm";
 }
 
 async function doesNpmAutoInstallPeerDependencies() {
@@ -496,6 +490,10 @@ async function getRecommendedDependenciesInstallationCommand(
     return ["yarn", "add", "--dev", ...deps];
   }
 
+  if (await isPnpmProject()) {
+    return ["pnpm", "add", "-D", ...deps];
+  }
+
   return ["npm", "install", "--save-dev", ...deps];
 }
 
@@ -503,7 +501,9 @@ async function getDependencies(
   projectType: SampleProjectTypeCreationAction
 ): Promise<Dependencies> {
   const shouldInstallPeerDependencies =
-    (await isYarnProject()) || !(await doesNpmAutoInstallPeerDependencies());
+    (await isYarnProject()) ||
+    (await isPnpmProject()) ||
+    !(await doesNpmAutoInstallPeerDependencies());
 
   const shouldInstallTypescriptDependencies =
     projectType === Action.CREATE_TYPESCRIPT_PROJECT_ACTION;